+++ src/command_modules/azure-cli-servicebus/azure/cli/command_modules/servicebus/_exception_handler.py
@@ -7,8 +7,9 @@
 def response_exception_handler(ex):
     from azure.mgmt.servicebus.models import ErrorResponseException
     from knack.util import CLIError
+    import json
 
-    if isinstance(ex, ErrorResponseException):
+    if isinstance(ex, ErrorResponseException) and ex.response.json():
         error_message = ex.response.json()
         error_message = {k.lower(): v for k, v in error_message.items()}
         if 'error' in error_message:
@@ -16,7 +17,10 @@ def response_exception_handler(ex):
         if 'code' in error_message and 'message' in error_message:
             message = '{}.'.format(error_message['message']) if error_message['message'] else 'Operation failed.'
             code = '[Code: "{}"]'.format(error_message['code']) if error_message['code'] else ''
-            raise CLIError('{} - {}'.format(code, message))
+            if code == '':
+                raise CLIError('Error Message - {}'.format(message))
+            else:
+                raise CLIError('{} - {}'.format(code, message))
         else:
             raise CLIError(ex)
     else: