+++ src/command_modules/azure-cli-servicebus/azure/cli/command_modules/servicebus/tests/latest/test_servicebus_commands.py
@@ -8,6 +8,7 @@
 import time
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, live_only)
 from azure.mgmt.servicebus.models import ProvisioningStateDR
+from knack.util import CLIError
 
 
 # pylint: disable=line-too-long
@@ -47,6 +48,14 @@ class SBNamespaceCURDScenarioTest(ScenarioTest):
         self.cmd('servicebus namespace show --resource-group {rg} --name {namespacename}',
                  checks=[self.check('sku.name', self.kwargs['sku'])])
 
+        try:
+            # Create Namespace with invalid name
+            self.cmd(
+                'servicebus namespace create --resource-group {rg} --name 1122334455 --location {loc} --tags {tags} --sku {sku}',
+                checks=[self.check('sku.name', self.kwargs['sku'])])
+        except CLIError as ex:
+            self.assertTrue(str(ex).find('The specified service namespace is invalid.'))
+
         # Update Namespace
         self.cmd(
             'servicebus namespace update --resource-group {rg} --name {namespacename} --tags {tags}',
@@ -79,6 +88,14 @@ class SBNamespaceCURDScenarioTest(ScenarioTest):
         self.cmd(
             'servicebus namespace authorization-rule show --resource-group {rg} --namespace-name {namespacename} --name {defaultauthorizationrule}',
             checks=[self.check('name', self.kwargs['defaultauthorizationrule'])])
+
+        try:
+            # List Authorizationrule with non-existence namespace
+            self.cmd(
+                'az servicebus namespace authorization-rule list --resource-group {rg} --namespace-name clitesting11')
+        except CLIError as ex:
+            self.assertTrue(str(ex).find('Can not perform requested operation on nested resource.'))
+
 
         # Get Authorization Rule Listkeys
         self.cmd(